Integrated Stripe Setup with Foundation Dashboard
